functions moved
