+++ django/chat/models.py
@@ -13,15 +13,16 @@ from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
 from data_fetcher.util import get_request
+from langchain_text_splitters import RecursiveCharacterTextSplitter
+from rapidfuzz import fuzz
 from structlog import get_logger
 
+from chat.llm import OttoLLM
 from chat.prompts import current_time_prompt
 from librarian.models import DataSource, Library, SavedFile
 from librarian.utils.process_engine import guess_content_type
 from otto.models import SecurityLabel, User
 from otto.utils.common import display_cad_cost, set_costs
-
-from .views import extract_claims_from_llm, highlight_claims
 
 logger = get_logger(__name__)
 
@@ -556,6 +557,54 @@ class AnswerSource(models.Model):
                     # return row[0]
                     return self.highlighted_text
         return _("Source not available (document deleted or modified since message)")
+
+    def highlight_claims(claims_list, text, threshold=80):
+        # match if the claims_list exist is text; if it does, then highlight it with  <mark> tag
+        text_splitter = RecursiveCharacterTextSplitter(
+            chunk_size=100,
+            chunk_overlap=20,
+            length_function=len,
+            is_separator_regex=False,
+        )
+
+        good_matches = []
+        # Split the text into chunks
+        chunks = text_splitter.create_documents([text])
+
+        for claim in claims_list:
+            for chunk in chunks:
+                chunk_text = chunk.page_content
+                # Find fuzzy matches
+                score = fuzz.partial_ratio(chunk_text.lower(), claim.lower())
+                if score >= threshold:
+                    good_matches.append(chunk_text)
+
+        for match in good_matches:
+            if len(match) > 3:
+                text = text.replace(match, f"<mark>{match}</mark>")
+
+        return text
+
+    def extract_claims_from_llm(llm_response_text):
+        llm = OttoLLM()
+        prompt = f"""
+        Based on the following LLM response, extract key factual claims, including direct quotes.
+
+        Respond in the format:
+        <claim>whatever the claim is...</claim>
+        <claim>another claim...</claim>
+
+        etc.
+
+        ---
+        <llm_response>
+        {llm_response_text}
+        </llm_response>
+        """
+        claims_response = llm.complete(prompt)
+        # find the claim tags and add whats wrapped in the claim tags to a list
+        claims_list = re.findall(r"<claim>(.*?)</claim>", claims_response)
+        return claims_list
 
     def update_claims_list(self, llm_response_text):
         """
