+++ django/chat/views.py
@@ -17,8 +17,6 @@ from django.utils.translation import get_language
 from django.utils.translation import gettext as _
 from django.views.decorators.http import require_GET, require_POST
 
-from langchain_text_splitters import RecursiveCharacterTextSplitter
-from rapidfuzz import fuzz
 from rules.contrib.views import objectgetter
 from structlog import get_logger
 from structlog.contextvars import bind_contextvars
@@ -942,53 +940,3 @@ def email_author(request, chat_id):
     )
     mailto_link = generate_mailto(to=chat.user.email, subject=subject, body=body)
     return HttpResponse(f"<a href='{mailto_link}'>mailto link</a>")
-
-
-def highlight_claims(claims_list, text, threshold=80):
-    # match if the claims_list exist is text; if it does, then highlight it with  <mark> tag
-    text_splitter = RecursiveCharacterTextSplitter(
-        chunk_size=100,
-        chunk_overlap=20,
-        length_function=len,
-        is_separator_regex=False,
-    )
-
-    good_matches = []
-    # Split the text into chunks
-    chunks = text_splitter.create_documents([text])
-
-    for claim in claims_list:
-        for chunk in chunks:
-            chunk_text = chunk.page_content
-            # Find fuzzy matches
-            score = fuzz.partial_ratio(chunk_text.lower(), claim.lower())
-            if score >= threshold:
-                good_matches.append(chunk_text)
-
-    for match in good_matches:
-        if len(match) > 3:
-            text = text.replace(match, f"<mark>{match}</mark>")
-
-    return text
-
-
-def extract_claims_from_llm(llm_response_text):
-    llm = OttoLLM()
-    prompt = f"""
-    Based on the following LLM response, extract key factual claims, including direct quotes.
-
-    Respond in the format:
-    <claim>whatever the claim is...</claim>
-    <claim>another claim...</claim>
-
-    etc.
-
-    ---
-    <llm_response>
-    {llm_response_text}
-    </llm_response>
-    """
-    claims_response = llm.complete(prompt)
-    # find the claim tags and add whats wrapped in the claim tags to a list
-    claims_list = re.findall(r"<claim>(.*?)</claim>", claims_response)
-    return claims_list
